+++ cmd/utils.go
@@ -10,11 +10,11 @@ import (
 	"strconv"
 	"strings"
 
-	"github.com/asdine/storm/v3"
 	"github.com/samber/lo"
 	"github.com/spf13/cobra"
 	"github.com/spf13/pflag"
 	"github.com/spf13/viper"
+	boltapi "go.etcd.io/bbolt"
 	yaml "gopkg.in/yaml.v3"
 
 	"github.com/filebrowser/filebrowser/v2/settings"
@@ -172,7 +172,7 @@ func withViperAndStore(fn func(cmd *cobra.Command, args []string, v *viper.Viper
 
 		log.Println("Using database: " + path)
 
-		db, err := storm.Open(path, storm.BoltOptions(databasePermissions, nil))
+		db, err := boltapi.Open(path, databasePermissions, nil)
 		if err != nil {
 			return err
 		}