+++ storage/bolt/bolt.go
@@ -1,7 +1,7 @@
 package bolt
 
 import (
-	"github.com/asdine/storm/v3"
+	boltapi "go.etcd.io/bbolt"
 
 	"github.com/filebrowser/filebrowser/v2/auth"
 	"github.com/filebrowser/filebrowser/v2/sessions"
@@ -12,14 +12,14 @@ import (
 )
 
 // NewStorage creates a storage.Storage based on Bolt DB.
-func NewStorage(db *storm.DB) (*storage.Storage, error) {
-	userStore := users.NewStorage(usersBackend{db: db.Bolt})
-	shareStore := share.NewStorage(shareBackend{db: db.Bolt})
-	settingsStore := settings.NewStorage(settingsBackend{db: db.Bolt})
-	authStore := auth.NewStorage(authBackend{db: db.Bolt}, userStore)
-	sessionStore := sessions.NewStorage(sessionBackend{db: db.Bolt})
+func NewStorage(db *boltapi.DB) (*storage.Storage, error) {
+	userStore := users.NewStorage(usersBackend{db: db})
+	shareStore := share.NewStorage(shareBackend{db: db})
+	settingsStore := settings.NewStorage(settingsBackend{db: db})
+	authStore := auth.NewStorage(authBackend{db: db}, userStore)
+	sessionStore := sessions.NewStorage(sessionBackend{db: db})
 
-	err := kvPut(db.Bolt, "version", 2)
+	err := kvPut(db, "version", 2)
 	if err != nil {
 		return nil, err
 	}